Prepare architecture changes
